Create Program to convert Celsius to Fahrenheit
This program is used to convert temperature from Celsius to Fahrenheit. Here also we've imported the Scanner through which we can receive input from the user,  ie. For Celsius. To convert and return value in Fahrenheit, 
Here we have used the Formulae for conversion of Celsius to Fahrenheit is
F = (c × 9/5) +32